requirements.txt
Flask==3.0.0

Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
python-dotenv==1.0.0
Pillow==10.0.0
requests==2.31.0